requirements.txt
flask
transformers
torch
python-docx
pdfplumber